+++ lizard_graph/urls.py
@@ -8,7 +8,6 @@ from django.views.generic import TemplateView
 from lizard_ui.urls import debugmode_urlpatterns
 
 from lizard_graph.views import GraphView
-from lizard_graph.views import HorizontalBarGraphView
 
 
 admin.autodiscover()
@@ -18,9 +17,6 @@ urlpatterns = patterns(
     url(r'^$',
         GraphView.as_view(),
         name="lizard_graph_graph_view"),
-    url(r'^bar/$',
-        HorizontalBarGraphView.as_view(),
-        name="lizard_graph_horizontal_bar_graph_view"),
     url(r'^examples/$',
         TemplateView.as_view(template_name="lizard_graph/examples.html")),
     )